backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
pip==23.2.1
setuptools==68.0.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
pillow==10.0.0